Corrected database connecting errors for gala pages

+++ galas/view.php
@@ -190,8 +190,9 @@
                 if (isset($_SESSION['username']) && (galaFullAccess($connection,$currentUser,$memberValidation))) {
                     echo linkButton("Add or Edit Gala Events", $domain . "gala-events.php?id=".$gala->getID(), true) . '<br>';
                 }
+
                 echo '</div><div class="large-4 medium-4 small-12 columns left">';
-                if (isset($_SESSION['username']) && (galaFullAccess($connection,$currentUser,$memberValidation))) {
+                if (isset($_SESSION['username']) && (galaFullAccess($connection,$currentUser,$memberValidation)) && $gala->doEventsExist($conn)) {
                     echo linkButton("Add Swim Times", $domain . "swim-times/create.php?galaID=".$gala->getID(), true) . '<br>';
                 }
                 echo '</div></div>';
